ISSUE #3008 - make avatar component independent
