+++ frontend/src/v5/ui/components/shared/userMenu/editProfileModal/editProfilePersonalTab/editProfileAvatar/editProfileAvatar.styles.ts
